# specs/openapi.yaml
openapi: 3.0.0
info:
  title: Simple-Sync API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Simple-Sync is an event storage and ACL-based access control service. This
    OpenAPI specification describes the public API surface implemented by the
    server and mirrors the documentation in `docs/src/content/docs/api/v1.md`.
paths:
...
  /user/generateToken:
    post:
      tags: [user]
      operationId: generateSetupToken
      summary: Generate a setup token for a user
      description: Generate a short-lived setup token which can be exchanged for
        an API key. Caller must be authenticated and have `.user.generateToken`
        permission for the target user.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateTokenRequest'
            examples:
              sample:
                value:
                  user: "user.123"
      responses:
        '200':
          description: Setup token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /user/exchangeToken:
    post:
      tags: [user]
      operationId: exchangeSetupToken
      summary: Exchange a setup token for an API key
      description: Exchange a previously generated setup token for a new API key.
        This endpoint is intentionally unauthenticated (used by initial setup
        flows). The response contains the generated API key (plain text) once.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                description:
                  type: string
              required:
                - token
      responses:
        '200':
          description: API key created from setup token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '406':
          $ref: '#/components/responses/NotAcceptable'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Use the `X-API-Key` header to pass a user's API key
  schemas:
...
    GenerateTokenRequest:
      type: object
      required:
        - user
      properties:
        user:
          type: string
          description: Target user identifier to generate a setup token for
      example:
        user: "user.123"
    SetupTokenResponse:
      type: object
      properties:
        token:
          type: string
          description: Setup token string
        expiresAt:
          type: string
          format: date-time
          description: Expiration timestamp of the setup token
      example:
        token: "ABCD-1234"
        expiresAt: '2025-09-26T12:00:00Z'
    ExchangeTokenResponse:
      type: object
      properties:
        keyUuid:
          type: string
          format: uuid
          description: UUID for the created API key record
        apiKey:
          type: string
          description: Plaintext API key value (only returned once)
        user:
          type: string
        description:
          type: string
          description: Optional human-friendly description for the key
      example:
        keyUuid: "0199ab65-1a1e-7000-80f5-23a591c5106e"
        apiKey: "sk_abcdefghijklmnopqrstuvwxyz1234567890"
        user: "user.123"
        description: "Desktop Client"
  responses:
    BadRequest:
      description: Bad request due to invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotAcceptable:
      description: Required header or media type not acceptable
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    MethodNotAllowed:
      description: HTTP method not allowed for this path
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
security:
  - ApiKeyAuth: []
...